Remove mention of Datacamp
We condemn in absolute terms the actions of individuals associated with Datacamp. Following the lead of RStudio (https://twitter.com/rstudio/status/1117889813879746560) and many other organizations, we have removed links and mentions of Datacamp. We stand opposed to harassment or assault of any kind, in or out of the workplace. We regret that this resource was not updated in a more timely manner.